Added <p> color and  text align in Basic_CSS_1.css; removed style attributes from <div> tags in Basic_HTML_3.html
I added color and text alignment for the <p> tags in Basic_CSS_1.css. This overrode the style in my <div> tags in Basic_HTML_3.html, so I removed the style attribute after trying and failing to find a solution that would allow me to keep the style for the three middle paragraphs and have CSS control the first and last paragraph. I'm sure as I learn more, I'll better understand how to accomplish this. This was done for the Linking CSS Assignment in the HTML and CSS course.