Solved color problem with class attribute in <p>
Added .font_1, .font_2, .font_3 to Basic_CSS_1.css. I set them as class attributes in the middle three paragraphs in Basic_HTML_3.hmtl, respectively. This solved my problem with losing the color for these paragraphs in the <div> style attributes because CSS was overriding it. I'm quite happy with the result, despite this not being for any actual assignment; I was just not happy with losing the color I'd set for these paragraphs.