Updated academy_cinemas.html
I added pill badges to each of the cards for the showtimes. The assignments said to put the card deck under the jumbotron piece of my site; however, the videos had it under the About section - they even put comments in for the showtimes in that place before the assignments. I personally think it looks better where the videos have it so I moved the entire section to under the About section.