Created second_challenge.html and second_challenge.css
I created second_challenge.html and second_challenge.css. second_challenge.css centers the footer element and makes it bold. I wasn't sure if using a font-weight number in CSS would be acceptable as compared to just saying "bold" so I played it safe.  This was for the Second Challenge in the Challenge Series in the HTML and CSS course.